correction

+++ ft_bzero.c
@@ -6,13 +6,15 @@
 /*   By: kyacini <kyacini@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/02 16:04:21 by kyacini           #+#    #+#             */
-/*   Updated: 2022/05/02 16:14:24 by kyacini          ###   ########.fr       */
+/*   Updated: 2022/05/03 23:56:35 by kyacini          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <stddef.h>
+
 void	bzero(void *s, size_t n)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*new;
 
 	new = s;
