+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: kyacini <kyacini@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/02 16:08:07 by kyacini           #+#    #+#             */
-/*   Updated: 2022/05/02 18:12:13 by kyacini          ###   ########.fr       */
+/*   Updated: 2022/05/03 23:52:50 by kyacini          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 void	*ft_memcpy(void *destination, const void *source, size_t size)
 {
-	int					i;
-	unsigned char		*new;
-	const unsigned char	*renew;
+	size_t					i;
+	unsigned char			*new;
+	const unsigned char		*renew;
 
 	new = destination;
 	renew = source;
