+++ ft_memset.c
@@ -6,13 +6,15 @@
 /*   By: kyacini <kyacini@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/02 15:48:23 by kyacini           #+#    #+#             */
-/*   Updated: 2022/05/02 16:01:35 by kyacini          ###   ########.fr       */
+/*   Updated: 2022/05/03 21:43:26 by kyacini          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <stddef.h>
+
 void	*memset(void *pointer, int value, size_t count)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*new;
 
 	new = pointer;
